# Remove space leak and inefficiencies associated with the Airship-Trace header.