This PR contains three optimizations to Airship tracing, which (on my 2014 MBP) yield a 18% gain in requests per second (from ~17000 to ~20000).

Before:
```
Running 30s test @ http://localhost:3000/
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   480.45us  272.68us   9.77ms   92.86%
    Req/Sec    16.80k     1.66k   19.19k    73.67%
  501314 requests in 30.00s, 181.53MB read
Requests/sec:  16709.68
Transfer/sec:      6.05MB
```

After:
```
Running 30s test @ http://localhost:3000/
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   386.71us  187.12us   9.02ms   89.36%
    Req/Sec    20.46k     1.21k   21.62k    90.03%
  612854 requests in 30.10s, 221.91MB read
Requests/sec:  20360.40
Transfer/sec:      7.37MB
```

The three optimizations are as follows:
1) We stop using `RWST`, as its `Writer` component leaks space. We replace it with an `RST` monad, extracted from the Snap framework, that provides assured strictness and an inlined `>>=` implementation. In addition, we use `modify'` rather than `modify` to ensure that the change is applied strictly (this yielded a small but significant increase in throughput).
2) We use `(:)` to add a new item onto the internal list of traces, rather than constructing a new list out of the item and using `mappend`. Though this constructs the list of traces in reverse order, we `reverse` said list before sending it to the client.
3) We use `ByteString` values in the internal trace list so as to avoid the overhead of an unnecessary `decodeUtf8` (since the header value is a `ByteString` anyway).

This significantly reduces the time we spend in `Airship.Internal.Decision.trace`. The code currently in master spends between 5-8% of its time in `trace`, as the following cost-center profile shows:

```
COST CENTRE              MODULE                        %time %alloc

utcTimeToRfc1123         Airship.Internal.Date           9.2    6.5
>>=                      Control.Monad.Trans.Either      7.0    5.0
trace                    Airship.Internal.Decision       5.5    9.9
lift                     Control.Monad.Trans.Either      5.3    7.5
o18                      Airship.Internal.Decision       4.6    6.2
fmap                     Airship.Types                   4.5    6.7
[ omitted ...]
```

After this optimization, it spends less than half that time in `trace`:
```
COST CENTRE              MODULE                        %time %alloc

>>=                      Control.Monad.Trans.Either     21.1   17.6
utcTimeToRfc1123         Airship.Internal.Date           7.9    7.8
lift                     Control.Monad.Trans.Either      7.8    6.8
state                    Airship.Types                   4.9    8.4
socketConnection.sendall Network.Wai.Handler.Warp.Run    3.4    0.1
return.\                 Airship.RST                     2.7    6.8
trace                    Airship.Internal.Decision       2.2    4.1
fmap                     Control.Monad.Trans.Either      2.0    1.9
```

(The presence of `return` in the above trace is a red herring: the original implementation would have a similarly significant chunk of `RWST` `return`s, but I compiled without library profiling so as to provide more readable profiles.)

These tests were performed on the `basic` application provided in the Airship distribution, with the `wrk` tool: `wrk -t1 -c8 -d30s http://localhost:3000/`. The application itself was run with three cores, using `stack exec basic -- +RTS -N3 -RTS`. The memory footprint of the `basic` application remains constant at ~8MB.