@patrickt 2016-07-28
@charleso Should be fixed.