@patrickt 2016-07-28
Closing to open this in a forked branch rather than my `master.